# Backend/app/Http/Requests/UpdateUser.php
<?php

namespace App\Http\Requests;

use Illuminate\Foundation\Http\FormRequest;

class UpdateUser extends FormRequest
{
    /**
     * Determine if the user is authorized to make this request.
     *
     * @return bool
     */
    public function authorize()
    {
        return true;
    }

    /**
     * Get the validation rules that apply to the request.
     *
     * @return array
     */
    public function rules()
    {
        return [
            'name'   => 'required|string|max:50',
            'thumbnail'    => 'nullable|image|mimes:jpg,jpeg,png|max:2000',
            //'email'       => 'required|string|email|max:155|unique:users',
        ];
    }
}