# Backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.2",
    "info": {
        "title": "Simple TODO  API Documentation",
        "description": "Service Documentation",
        "contact": {
            "email": "info@std.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/todo/v1",
            "description": "STD API Server"
        }
    ],
    "paths": {
...
        "/task": {
            "get": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Authority: User | Get all Tasks",
                "description": "This endpoint retrieves all tasks",
                "operationId": "taskIndex",
                "responses": {
                    "200": {
                        "description": "Tasks retrieved",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Task"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Authority: User | Create a new Task",
                "description": "This endpoint creates a new Task.",
                "operationId": "createTask",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateTask"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Task created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Task"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized. User not with access role"
                    },
                    "422": {
                        "description": "The given data was invalid."
                    }
                }
            }
        },
        "/task/{taskId}": {
            "get": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Authority: User | Get details of a Task by ID",
                "description": "This endpoint returns all related data of specified task",
                "operationId": "getTask",
                "responses": {
                    "200": {
                        "description": "Task retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Task"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated. Token needed"
                    },
                    "403": {
                        "description": "Unauthorized. User not with access role"
                    },
                    "404": {
                        "description": "Resource not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Authority: User | Update a Task | Please use x-www-form-urlencoded for body",
                "description": "This endpoint updates the task",
                "operationId": "updateTask",
                "parameters": [
                    {
                        "name": "taskId",
                        "in": "path",
                        "description": "Task ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateTask"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Task updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Task"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized. User not with access role"
                    },
                    "404": {
                        "description": "Task not found"
                    },
                    "422": {
                        "description": "The given data was invalid."
                    }
                }
            },
            "delete": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Authority: User | Delete a Task",
                "description": "This endpoint deletes the selected Task",
                "operationId": "deleteTask",
                "parameters": [
                    {
                        "name": "taskId",
                        "in": "path",
                        "description": "Task ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Task deleted successfully"
                    },
                    "403": {
                        "description": "Unauthorized. User not with access role"
                    },
                    "404": {
                        "description": "Task not found"
                    },
                    "422": {
                        "description": "The given data was invalid."
                    }
                }
            }
        },
...
        "/updateProfile": {
            "put": {
                "tags": [
                    "Users"
                ],
                "summary": "Authority: User | Update authenticated User | Please use x-www-form-urlencoded for body",
                "description": "This endpoint updates the current user",
                "operationId": "updateUser",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateUser"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized. User not with access role"
                    },
                    "404": {
                        "description": "User not found"
                    },
                    "422": {
                        "description": "The given data was invalid."
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "Task": {
                "title": "Task",
                "description": "Task model",
                "properties": {
                    "id": {
                        "title": "ID",
                        "description": "ID",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "title": {
                        "title": "title",
                        "description": "The title of the Task",
                        "type": "string",
                        "example": "Lead a Series E"
                    },
                    "description": {
                        "title": "description",
                        "description": "The description of the Task",
                        "type": "string",
                        "example": "Becoming a Unicorn"
                    },
                    "thumbnail": {
                        "title": "thumbnail",
                        "description": "The image of the Task",
                        "type": "string",
                        "example": "https://res.cloudinary.com/dtgglytjm/image/upload/v1612307531/Files/i9gts5zn0wiusxekrrjz.png"
                    },
                    "priority": {
                        "title": "priority",
                        "description": "The priority level of the Task",
                        "type": "string",
                        "example": "high"
                    },
                    "start_date": {
                        "title": "start_date",
                        "description": "Start date of Task",
                        "type": "string",
                        "format": "datetime",
                        "example": "2020-01-27 17:50:45"
                    },
                    "end_date": {
                        "title": "end_date",
                        "description": "End date of Task",
                        "type": "string",
                        "format": "datetime",
                        "example": "2020-01-27 17:50:45"
                    },
                    "access": {
                        "title": "access",
                        "description": "The access type of the Task",
                        "type": "boolean",
                        "example": "0"
                    },
                    "status": {
                        "title": "status",
                        "description": "The status of the Task",
                        "type": "string",
                        "example": "completed"
                    },
                    "created_at": {
                        "title": "Created at",
                        "description": "Created at",
                        "type": "string",
                        "format": "datetime",
                        "example": "2021-09-27 17:50:45"
                    },
                    "updated_at": {
                        "title": "Updated at",
                        "description": "Updated at",
                        "type": "string",
                        "format": "datetime",
                        "example": "2021-09-27 17:50:45"
                    },
                    "user_id": {
                        "title": "user_id",
                        "description": "User ID",
                        "type": "integer",
                        "example": 1
                    },
                    "user": {
                        "$ref": "#/components/schemas/User"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Task"
                }
            },
            "User": {
                "title": "User",
                "description": "User model",
                "properties": {
                    "id": {
                        "title": "ID",
                        "description": "ID",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "name": {
                        "title": "name",
                        "description": "The name of the User",
                        "type": "string",
                        "example": "Peter Oduntan"
                    },
                    "email": {
                        "title": "email",
                        "description": "The email of the User",
                        "type": "string",
                        "example": "peter_oduntan@yahoo.com"
                    },
                    "thumbnail": {
                        "title": "thumbnail",
                        "description": "The image of the User",
                        "type": "string",
                        "example": "https://example.com/image.png"
                    },
                    "created_at": {
                        "title": "Created at",
                        "description": "Created at",
                        "type": "string",
                        "format": "datetime",
                        "example": "2021-09-27 17:50:45"
                    },
                    "updated_at": {
                        "title": "Updated at",
                        "description": "Updated at",
                        "type": "string",
                        "format": "datetime",
                        "example": "2021-09-27 17:50:45"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "User"
                }
            },
...
            "CreateTask": {
                "title": "Create Task Request",
                "description": "Body Data",
                "required": [
                    "title",
                    "description",
                    "thumbnail",
                    "priority",
                    "start-date",
                    "end_date",
                    "access",
                    "status"
                ],
                "properties": {
                    "title": {
                        "title": "title",
                        "description": "The title of the Task",
                        "type": "string",
                        "example": "Doing Laundry"
                    },
                    "description": {
                        "title": "description",
                        "description": "The description of the task created",
                        "type": "string",
                        "example": "Going to the Laundry mart"
                    },
                    "thumbnail": {
                        "title": "thumbnail",
                        "description": "The image of the Task",
                        "type": "string",
                        "example": "https://en.wikipedia.org/image.png"
                    },
                    "priority": {
                        "title": "priority",
                        "description": "The priority level of the Task",
                        "type": "string",
                        "example": "low"
                    },
                    "start_date": {
                        "title": "start_date",
                        "description": "Start date of Task",
                        "type": "string",
                        "format": "datetime",
                        "example": "2020-01-27 17:50:45"
                    },
                    "end_date": {
                        "title": "end_date",
                        "description": "End date of Task",
                        "type": "string",
                        "format": "datetime",
                        "example": "2020-01-27 17:50:45"
                    },
                    "access": {
                        "title": "access",
                        "description": "The access type of the Task",
                        "type": "string",
                        "example": 1
                    },
                    "status": {
                        "title": "status",
                        "description": "The status of the Task",
                        "type": "string",
                        "example": "completed"
                    }
                },
                "type": "object"
            },
...
            "UpdateTask": {
                "title": "Create Task Request",
                "description": "Body Data",
                "required": [
                    "title",
                    "description",
                    "thumbnail",
                    "priority",
                    "start-date",
                    "end_date",
                    "access",
                    "status"
                ],
                "properties": {
                    "title": {
                        "title": "title",
                        "description": "The title of the Task",
                        "type": "string",
                        "example": "Doing Laundry"
                    },
                    "description": {
                        "title": "description",
                        "description": "The description of the task created",
                        "type": "string",
                        "example": "Going to the Laundry mart"
                    },
                    "thumbnail": {
                        "title": "thumbnail",
                        "description": "The image of the Task",
                        "type": "string",
                        "example": "https://en.wikipedia.org/image.png"
                    },
                    "priority": {
                        "title": "priority",
                        "description": "The priority level of the Task",
                        "type": "string",
                        "example": "low"
                    },
                    "start_date": {
                        "title": "start_date",
                        "description": "Start date of Task",
                        "type": "string",
                        "format": "datetime",
                        "example": "2020-01-27 17:50:45"
                    },
                    "end_date": {
                        "title": "end_date",
                        "description": "End date of Task",
                        "type": "string",
                        "format": "datetime",
                        "example": "2020-01-27 17:50:45"
                    },
                    "access": {
                        "title": "access",
                        "description": "The access type of the Task",
                        "type": "string",
                        "example": 1
                    },
                    "status": {
                        "title": "status",
                        "description": "The status of the Task",
                        "type": "string",
                        "example": "completed"
                    }
                },
                "type": "object"
            },
            "UpdateUser": {
                "title": "User Creation Request",
                "description": "User Request Body Data",
                "required": [
                    "name",
                    "email",
                    "password",
                    "thumbnail"
                ],
                "properties": {
                    "name": {
                        "title": "name",
                        "description": "The name of the user",
                        "type": "string",
                        "example": "Peter"
                    },
                    "email": {
                        "title": "email",
                        "description": "The email of the user",
                        "type": "string",
                        "example": "peter_oduntan@yahoo.com"
                    },
                    "password": {
                        "title": "password",
                        "description": "The password of the user",
                        "type": "string",
                        "example": "Secret123"
                    },
                    "thumbnail": {
                        "title": "thumbnail",
                        "description": "The image of the User",
                        "type": "string",
                        "example": "https://example.com/image.png"
                    }
                },
                "type": "object"
            },
...
        }
    },
    "security": [
        []
    ]
}